Python bindings: Extract MPI adaptor to Mpi.hpp
