+++ src/binding/python/Series.cpp
@@ -29,27 +29,12 @@
 #if openPMD_HAVE_MPI
 //  re-implemented signatures:
 //  include <mpi4py/mpi4py.h>
+#include "openPMD/binding/python/Mpi.hpp"
 #include <mpi.h>
 #endif
 
 #include <sstream>
 #include <string>
-
-#if openPMD_HAVE_MPI
-/** mpi4py communicator wrapper
- *
- * refs:
- * - https://github.com/mpi4py/mpi4py/blob/3.0.0/src/mpi4py/libmpi.pxd#L35-L36
- * - https://github.com/mpi4py/mpi4py/blob/3.0.0/src/mpi4py/MPI.pxd#L100-L105
- * - installed: include/mpi4py/mpi4py.MPI.h
- */
-struct openPMD_PyMPICommObject
-{
-    PyObject_HEAD MPI_Comm ob_mpi;
-    unsigned int flags;
-};
-using openPMD_PyMPIIntracommObject = openPMD_PyMPICommObject;
-#endif
 
 struct SeriesIteratorPythonAdaptor : SeriesIterator
 {
@@ -229,65 +214,17 @@ It will be replaced with an automatically determined file name extension:
                         Access at,
                         py::object &comm,
                         std::string const &options) {
-                //! TODO perform mpi4py import test and check min-version
-                //!       careful: double MPI_Init risk? only import mpi4py.MPI?
-                //!       required C-API init? probably just checks:
-                //! refs:
-                //! -
-                //! https://bitbucket.org/mpi4py/mpi4py/src/3.0.0/demo/wrap-c/helloworld.c
-                //! - installed: include/mpi4py/mpi4py.MPI_api.h
-                // if( import_mpi4py() < 0 ) { here be dragons }
-
-                if (comm.ptr() == Py_None)
-                    throw std::runtime_error(
-                        "Series: MPI communicator cannot be None.");
-                if (comm.ptr() == nullptr)
-                    throw std::runtime_error(
-                        "Series: MPI communicator is a nullptr.");
-
-                // check type string to see if this is mpi4py
-                //   __str__ (pretty)
-                //   __repr__ (unambiguous)
-                //   mpi4py: <mpi4py.MPI.Intracomm object at 0x7f998e6e28d0>
-                //   pyMPI:  ... (TODO)
-                py::str const comm_pystr = py::repr(comm);
-                std::string const comm_str = comm_pystr.cast<std::string>();
-                if (comm_str.substr(0, 12) != std::string("<mpi4py.MPI."))
-                    throw std::runtime_error(
-                        "Series: comm is not an mpi4py communicator: " +
-                        comm_str);
-                // only checks same layout, e.g. an `int` in `PyObject` could
-                // pass this
-                if (!py::isinstance<py::class_<openPMD_PyMPIIntracommObject> >(
-                        comm.get_type()))
-                    // TODO add mpi4py version from above import check to error
-                    // message
-                    throw std::runtime_error(
-                        "Series: comm has unexpected type layout in " +
-                        comm_str +
-                        " (Mismatched MPI at compile vs. runtime? "
-                        "Breaking mpi4py release?)");
-
-                // todo other possible implementations:
-                // - pyMPI (inactive since 2008?): import mpi; mpi.WORLD
-
-                // reimplementation of mpi4py's:
-                // MPI_Comm* mpiCommPtr = PyMPIComm_Get(comm.ptr());
-                MPI_Comm *mpiCommPtr =
-                    &((openPMD_PyMPIIntracommObject *)(comm.ptr()))->ob_mpi;
-
-                if (PyErr_Occurred())
-                    throw std::runtime_error(
-                        "Series: MPI communicator access error.");
-                if (mpiCommPtr == nullptr)
+                auto variant = pythonObjectAsMpiComm(comm);
+                if (auto errorMsg = std::get_if<std::string>(&variant))
                 {
-                    throw std::runtime_error(
-                        "Series: MPI communicator cast failed. "
-                        "(Mismatched MPI at compile vs. runtime?)");
+                    throw std::runtime_error("[Series] " + *errorMsg);
                 }
-
-                py::gil_scoped_release release;
-                return new Series(filepath, at, *mpiCommPtr, options);
+                else
+                {
+                    py::gil_scoped_release release;
+                    return new Series(
+                        filepath, at, std::get<MPI_Comm>(variant), options);
+                }
             }),
             py::arg("filepath"),
             py::arg("access"),
